2/27/26 - testsystem3

+++ dat_learning_system-be/LMS.Backend/DTOs/Classroom/ClassroomDto.cs
@@ -20,7 +20,7 @@ public class ClassroomLessonDto
     public int SortOrder { get; set; }
     public bool IsLocked { get; set; }
     public bool IsDone { get; set; }
-    public double? LastScore { get; set; }
+    public decimal? LastScore { get; set; }
     public List<ClassroomContentDto> Contents { get; set; } = new();
 }
